refactored inside_building/building_is_active/active_building
