+++ src/market.c
@@ -36,9 +36,9 @@ static unsigned int get_markets(region * r, unit ** results, size_t size)
     if (!btype)
         return 0;
     for (b = r->buildings; n < size && b; b = b->next) {
-        if (b->type == btype && (b->flags & BLD_WORKING)
-            && b->size >= b->type->maxsize) {
+        if (b->type == btype && building_is_active(b)) {
             unit *u = building_owner(b);
+            /* I decided to omit check for inside_building(u) */
             unsigned int i;
             for (i = 0; u && i != n; ++i) {
                 /* only one market per faction */
