+++ src/study.c
@@ -175,13 +175,6 @@ static int study_days(unit * student, skill_t sk)
     return student->number * speed;
 }
 
-static building *active_building(const unit *u, const struct building_type *btype) {
-    if (u->building && u->building->type == btype && building_is_active(u->building)) {
-        return inside_building(u);
-    }
-    return 0;
-}
-
 static int
 teach_unit(unit * teacher, unit * student, int nteaching, skill_t sk,
 bool report, int *academy)
@@ -546,8 +539,6 @@ int study_cmd(unit * u, order * ord)
     int maxalchemy = 0;
     int speed_rule = (study_rule_t)get_param_int(global.parameters, "study.speedup", 0);
     static int learn_newskills = -1;
-    struct building *b = inside_building(u);
-    const struct building_type *btype = building_is_active(b) ? b->type : NULL;
 
     if (learn_newskills < 0) {
         const char *str = get_param(global.parameters, "study.newskills");
@@ -610,10 +601,7 @@ int study_cmd(unit * u, order * ord)
         return 0;
     }
     /* Akademie: */
-    b = inside_building(u);
-    btype = building_is_active(b) ? b->type : NULL;
-
-    if (btype && btype == bt_find("academy")) {
+    if (active_building(u, bt_find("academy"))) {
         studycost = _max(50, studycost * 2);
     }
